Added pdf and thumbnail or large image functions

+++ Assignment_Two_2022/src/main/java/com/sd4/model/Beer.java
@@ -35,8 +35,8 @@ public class Beer extends RepresentationModel<Beer> implements Serializable {
     private long id;
     private long brewery_id;
     private String name;
-    private Integer cat_id;
-    private Integer style_id;
+    private long cat_id;
+    private long style_id;
     private Double abv;
     private Double ibu;
     private Double srm;
