+++ Assignment_Two_2022/src/main/java/com/sd4/model/Style.java
@@ -23,8 +23,8 @@ import lombok.ToString;
 public class Style implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-    private int cat_id;
+    private long id;
+    private long cat_id;
     private String style_name;
 
     @Temporal(TemporalType.TIMESTAMP)
